Moved declarations of methods to separate class
The Method class wen to long. The move of static declarations had made
it much more readable.
