+++ src/main/java/parser/MathExpression.java
@@ -5,6 +5,7 @@
 package parser;
 
 import interfaces.Savable;
+import parser.methods.Declarations;
 import parser.methods.Method;
 
 import math.Maths;
@@ -1314,7 +1315,6 @@ public class MathExpression implements Savable {
      * @return the result of the evaluation
      */
     public String solve() {
-
         if (correctFunction && !hasFunctionOrVariableInitStatement) {
             final ArrayList<String> myScan = new ArrayList<String>();
